Set lport name to be the port ID
This change partially reverts commit
fa1767eeb47a638622f631817ea2868022e5a1ea.

The reason is that the name must match the iface-id used for the
Interfaces table in the Open_vSwitch database.  This is spelled out in
the ovn design (see ovn/ovn-nb.xml) with the following text:

      <p>
      For entities (VMs or containers) that are spawned in the hypervisor,
      the name used here must match those used in the <ref key="iface-id"
      table="Interface" column="external_ids" db="Open_vSwitch"/> in the
      <ref db="Open_vSwitch"/> database's <ref table="Interface"
      db="Open_vSwitch"/> table, because hypervisors use <ref key="iface-id"
      table="Interface" column="external_ids" db="Open_vSwitch"/> as a lookup
      key to identify the network interface of that entity.
      </p>

However, the reason this was changed was because of some matching
behavior implemented in ovn-nbctl.  It's not possible to have a name
in the form of a UUID, as ovn-nbctl will only do a UUID based lookup.
ovn-nbctl will be fixed so that it will do a lookup with a UUID as a
name if a UUID based lookup fails.  Then we can merge this change.

Change-Id: Ib19bc52f992a25583abf995ce0119707f656fb11