todo: note neutronclient binding:profile patch
Add a reference to the python-neutronclient patch needed for sending a
custom binding:profile.

In passing, remove completed devstack task and update the description
of another task to not dictate the solution, because the solution is
likely to be just to install it from pypi once we get it posted there.

Change-Id: I4d2f6c444f2ff4156d01d0a91fee92b8a51e887b