devstack: Add multi-node support
Instead of having a single "ovn" service that can be enabled, this
patch makes it so you can enable "ovn-controller" and "ovn-northd"
separately.  This is important for multi-node devstack testing, as you
only want to run ovn-controller on a compute node.  Documentation and
sample configuration has been updated to reflect how to use these
updates.

This patch also adds two new devstack config variables: OVN_REMOTE and
OVN_UUID.

The default for OVN_REMOTE should be fine except on hosts where
ovn-controller is enabled.  In that case, it should be set such that
the host knows to talk to the proper host running ovsdb-server with
the OVN databases (the one running ovn-northd).

OVN_UUID is optional.  Normally a random UUID is generated (and saved
for re-use), but it can be specified if you desire to use a specific
one for some reason.

Change-Id: Iaf040fe17b26c9b7d0b61daa0bec1b3acbee3806