Control list of enabled services for tempest job.
This patch moves the list of enabled services into the local job
config so that we can more easily update it.  This list is the same as
what was in project-config with "n-obj" added.  One of the test
failures was
tempest.thirdparty.boto.test_ec2_instance_run.InstanceRunTest failing
when trying to create an S3 bucket, so I think turning this service
on will resolve that.

Change-Id: Ie9669381f0a1d5a695d998ab8453eb2427d39926
Signed-off-by: Russell Bryant <rbryant@redhat.com>