Convert plugin away from ML2
This patch converts the ovn plugin away from the ml2 plugin to be a
monolithic plugin. The reason is that we not really need the overhead and
complexity that ml2 adds and a monolithic plugin will serve us better.

Note: there are a lot of cleanups to do after this merges. Currently,
I'm skipping the existing unit tests that mocked the ovn interaction. I'll
add these back in a later patch set once we get this in.

Depends-On: I6ce60dad222e1e244f0d4fac9680e73de2a9b2e9
Change-Id: Ife1ac9910d4dd3af321586ec59c26df1d336d302