Fix OVN_PORT_BINDING_PROFILE
The conversion to the ml2 plugin broke this integration. This patch fixes it.
It also removes the function _validate_bindign_profile and instead uses
_get_data_from_binding_profile to return the data and validate it this way
we only need to call one method. Lastly, this patch changes the prototypes to
create_port_in_ovn to pass parent_name and tag as these are not in the port
object.

Change-Id: I89a6b732d5bfaf9625276caa6ec3ed51f6bf0416