devstack: Set VM MTU to 1400
It's easy to run into MTU problems when using OVN if the MTU in your VMs
is the same as the underlying network.  In that case, packets will exceed
the MTU when the tunnel headers get added.  Update the devstack plugin
to automatically avoid the most common instance of this problem by just
hard coding the VM MTU to 1400.

Change-Id: I8401af5b228eeb24505e72baeabf3d33f4991665
Signed-off-by: Russell Bryant <rbryant@redhat.com>