Add subtransactions=True on transaction for delete_port/network
These methods are sometimes called within a transcations so this is
needed. One specific case which has been throwing a trace in the
gate is during a race condition on delete_network/delete_port.
In this cause the dhcp-agent tries to delete it's port but this
is initiated via a base class and causes an error.

Change-Id: I910c1b05bcd5abd0582d10755deb40ddc4674234