Skip test_dualnet_multi_prefix_dhcpv6_stateless
This test seems to fail about 20% of the time in the gate because it's
unable to ping one of the v6 addresses on the instance. This doesn't
seem to be a neutron/ovn_plugin issue as far as I can tell. This patch
skips this test for now until we can tackle why this is failing.

Change-Id: I6a6958e6c424ad13027623f66e1190f8e887bbec
Related-bug: #1499024