Explicitly install tox
When testing OVN with Kuryr using a setup which only installs Neutron,
keystone, mysql and OVN, the devstack run fails because tox is not
installed. It was being pulled in when installing tempest. Since
networking-ovn uses genconfig to generate a config file, make sure
we install this ourselves.

With this run, my devstack with Kuryr works now.

Change-Id: I67afce78152f250920ef43334e7ee87440bcd075
Signed-off-by: Kyle Mestery <mestery@mestery.com>