Fix argument order in assertEqual to (expect, obs)
assertEqual expects that the arguments provided to it should be
(expected, observed). If a particular order is kept as a convention,
then it helps to provide a cleaner message to the developer if Unit Tests
fail. There are several Unit Test files where the arguments for assertEqual
have been swapped.

Change-Id: Idf99e2b43dad5ae2b7a0ed17ea6776253f823282
Related-Bug: #1259292