Update Vagrant configuration to deploy 3-node OVN setup
This updates the Vagrant setup to deploy a 3-node Ubuntu Trusty
setup (1 control node, 2 compute nodes). The Vivid image this
was previously trying to download was HUGE, and it's not clear
why running on Vivid is better considering the gate uses
Trusty.

Also, adjust the names of the nodes this spawns to be a bit
more sane. They are now spawned as:

ovn-controller
ovn-compute1
ovn-compute2

One final change adds some additional configuration into
local.conf which allows this all to come up and run.

Closes-Bug: #1499917
Change-Id: Ife6cfbe18ea1053ca52b8ff0f074987c178a713c
Signed-off-by: Kyle Mestery <mestery@mestery.com>