Prevent DBError caused by wrong param passed to get_security_group
Previously, in update_port() we would call _get_port() to get a copy
of the original port before we actually updated it. This original_port
value is in the form of a sqlalchemy object though and the method
update_security_group_on_port() uses the original port value to populate
update_port['security_groups'] I suppose to be helpful to the caller. In
our case, we'd later pass this value to get_security_groups and it would
give us an error of: DBError because we were passing a
neutron.db.securitygroups_db.SecurityGroupPortBinding rather than an id.

Change-Id: I04c522f6aed112dd3345f99ac80d582e667bab80
Closes-bug: #1514011