Add Rally to networking-ovn
This adds the needed infrastructure to connect networking-ovn
with Rally benchmarking.
A relevant job will be defined to run Neutron with OVN with
the ovn.yaml jobs.
It is easy to defined OVN specific plugins and change the
scenarios run by Rally for OVN using the added files in this
patch.

Change-Id: I7c6a518ba8fb01725cb7a32f5e62062afc16f632