Add helper method _drop_all_ip_traffic_to_port and tests
This patch simply pulls out some code from _add_acl to a helper method.
I'm starting to go through and add test coverage for the acl methods in
the plugin. Eventually my goal is to pull out all the acl helper methods
to their own class and file outside of the plugin.

Change-Id: If9fb28c7c7a0e23851d81db3ce241a0158d1d703