Support remote_group_id across networks by matching IP addresses
Our current code for handling a remote-group in a security group
rule is to list the OVN ports by name in the OVN ACL. This only
works for ports on the same OVN logical switch (Neutron network).
A remote group in a security group rule should work for ports on
other networks, as well.

This patch fixes the problem by matching on IP addresses rather
than OVN port names.

Closes-bug: 1525972

Change-Id: If69871389a9647e1e25a6cf5cdc76858fa1939db