Deployment: Update with OVN DB requirements
By default, the node running ovn-northd (and likely the central
OVN database) will not have the OVN NB and SB databases created,
nor set to load by the ovsdb-server instance on this host. This
patch adds some notes to the deployment scripts to ensure those
following this guide create those DBs and load them into the
ovsdb-server on startup.

Change-Id: I4cdd55e3910421921ddc4bd7beb2d9f1b769a6c4
Signed-off-by: Kyle Mestery <mestery@mestery.com>