doc: Further reflect that OVN L3 is the default.
The devstack testing docs still had a lot of text based on the Neutron
L3 agent being enabled.  Change the docs to focus on what happens with
OVN L3 is enabled, since that is the default.

This mainly involves dropping references to the "public" network, router
ports, and the use of floating IPs.

Change-Id: I08e42a02e3b8d6e7dfe7fac09632f190458fd50e
Signed-off-by: Russell Bryant <rbryant@redhat.com>