Update the HA section of the FAQ.
I realized that there were some important cavets in the HA approach of
letting Neutron rebuild the OVN northbound database on failure.  While
trying to document the caveats, I came to the conclusion that this is
not currently a practical enough approach to document.  If we improve
things, we can add it back to the document.

Change-Id: Ib5449661792d95fd72783f963ce8d32dfdf316c2
Signed-off-by: Russell Bryant <rbryant@redhat.com>