Resolve NetworkInUse race condition.
There is a race condition in delete_network() that we need to work
around here.  delete_network() issues a query to automatically delete
DHCP ports and then checks to see if any ports exist on the network.  If
a network is created and deleted quickly, such as when running tempest,
the DHCP agent may be creating its port for the network around the same
time that the network is deleted.  This can result in the DHCP port
getting created in between these two queries in delete_network().  To
work around that, we'll call delete_network() a second time if we get a
NetworkInUse exception but the only port(s) that exist are ones that
delete_network() is supposed to automatically delete.

Change-Id: I53f4d6fef94974d561cfe90665bfdca4392f62d9
Co-authored-by: Aaron Rosen <aaronorosen@gmail.com>