Add extension fields on network and port create
Add extension fields to the response on network and port create.
This patch set is based on [1] which calls for each core plugin
to apply similar logic to support extensions that rely on
extend_dict functions. For the OVN plugin, this applies to the
network_availability_zone extension for networks and the
extra_dhcp_opt extension for ports.

[1] https://review.openstack.org/#/c/276219/

Change-Id: I2fa69102ec3562a818fdf9b18f319da7981e4bf5
Closes-Bug: #1541774
Related-Bug: #1539274