Vagrant: Add live migration support
This patch adds the ability to live migrate guest VMs between compute
nodes in the Vagrant environment. This is done by setting up an NFS
server on the ovn-controller node, and mounting a shared nova instance
directory on compute nodes. We also tweak the libvirt configuration on
each host to listen on the appropriate sockets.

This patch also adds an SSH key pair for the Vagrant environment.

Change-Id: Ifdf24eebb2d2b7a5ba8080d3ae93c92ace5a0023
Signed-off-by: Kyle Mestery <mestery@mestery.com>
Co-Authored-By: Richard Theis <rtheis@us.ibm.com>