Replace infinite while loop with test_with_retry
Replace while loops in plugin.sh where devstack for ovn would
wait for a certain condition to be met. Using test_with_retry
from functions, we save a few lines and get a better sense of
how long the expected condition took to happen (via time_start)
as well as avoid the risk of waiting forever.

Change-Id: I213e7721c421b16093a7426c48c94e001ba745ff