Update supported API extensions for tempest.
I noticed in a tempest log that it thought security groups were
disabled.  This is because our list of enabled API extensions for the
tempest config was not up to date.  Sort the list in the plugin and
update the devstack plugin to always use the same list as the plugin.

Change-Id: I92ecb4b21a5581a925c164953df0710ead5eb402
Signed-off-by: Russell Bryant <rbryant@redhat.com>