Start ovn-northd in the devstack screen as a foreground process
A side-effect of separate ovsdb-server processors handling NB and SB
for default devstack (https://review.openstack.org/#/c/276364/) is
the use of ovn-ctl start_northd command to spawn the ovn-northd
process. The functionality of the devstack setup is correct, but in
the screen associated with the ovn-northd process, the output seen
is the following:

>> & echo $! >/opt/stack/status/stack/ovn-northd.pid; fg || \
     echo "ovn-northd failed to start" | \
     tee "/opt/stack/status/stack/ovn-northd.failure"
bash: syntax error near unexpected token `&'
>>

The ovn-northd process is actually spawned correctly and can be seen
in the output of ps. The start_northd creates ovn-northd as a daemon
and as a result the screen associated with ovn-northd doesn't have
a handle to the process.

This patch starts the ovn-northd as a foreground process in the
devstack screen. It also provides some simple command line options
associated with start_northd command in the documentation for a
non-devstack use case.

Change-Id: Ib5921fb7ba9efa0b5cf4cc8c99897473628dfae9