Create bridge for OVN L3 in devstack
When OVN is used for L3 routing in devstack, create the public
network, subnets, router, and also create the public bridge br-ex.
This is meant to be equivalent to what happens in devstack when the
L3 agent is enabled.  I hoped to do this by creating a flag that
could be set by plugins that would trigger the devstack code that
already exists to do this, but that approach was rejected (see
https://review.openstack.org/#/c/304871/).  So, the next best thing
is to copy that code from devstack into the networking-ovn devstack
plugin, which is what I did here.

Change-Id: I7e17b9f9f967a2625aa81b9dfb50739567cc223c
Closes-Bug: #1567068