Leverage common/acl in existing core_plugin
This patch removes the duplicated code from plugin.py thats in
common/acl.py. The main motivation behind this patch is to help
consolidate changes so that the ml2 plugin is using the same code
as the core_plugin which is being tested in the gate and removing
the code duplication.

Change-Id: I68c5a589a6f814152cc9b474efb5a5aa5759c240