Change override-defaults to not override Q_PLUGIN
Previously, if one wanted to try the ml2 plugin one would need to
manually modify this file. This is because it's source in the
plugin.sh file and overrides the previous values in the env.
This patch fixes this by only setting Q_PLUGIN to 'ovn' if it's not
already set.

Change-Id: I44cd0c967d7d12e9166d974cfa41413c12cff3a5