OVN NB sync doesn't ignore non-neutron routers
OVN NB sync will attempt to process routers not created by neutron.
This processing isn't necessary and could lead to OVN logical routers
or logical router ports being deleted. In practice, it is unlikely
that a delete will occur because the "neutron-" prefix would need
to be in the logical router name.

Change-Id: Ib353e6d158f948ba025c5b7ea57602ab76258232
Closes-Bug: #1586074