Add notes for ML2 driver port binding 500 errors
The neutron ML2 plugin does not provide drivers with an interface
to validate port bindings before the precommit stage. This makes it
harder for users to determine mistakes in the binding profile. This
problem resulted from the core plugin to ML2 driver conversion.

This patch set adds notes explaining the problem so it can be fixed
as soon as the neutron ML2 plugin provides such support.

Change-Id: I2944ddb2d753a708bbf623032e4043e9fbb44a57
Closes-Bug: #1588848