Add functional tests and gate hook functions
This patch adds few functional tests to test OVN NB DB sync
and adds the gate hook script functions.

The functional test setup
 - will start an ovsdb-server with OVN NB DB.
 - will start an ovsdb-server with OVN SB DB. This connection is
   not used by this patch. But it is created because
      - It can be used later to test the chassis up/down events.
      - ML2 OVN driver now also connects to the OVN SB DB.
 - The ML2 OVN driver is configured to connect to these ovsdb-servers
 - Another IDL connection is established with the OVN NB DB server.

The test cases are validated by checking if the OVN NB DB is
updated properly or not.

This can be extended later to add more functional tests and
full stack tests if required.

Change-Id: I4dc3c633e42776efc7b574f15c1bbd95f28f70f0