Add OVN L3 Router Scheduler
This patch provides the infrastructure for the OVN L3 to schedule
the router gateway port. The OVN hypervisor administrator can
set aside certain hypervisors to be used for scheduling router
gateway ports. This is the corresponding WIP patch at the moment:
http://openvswitch.org/pipermail/dev/2016-June/073285.html

For now it is assumed that any hypervisor/chassis can host a router
gateway port.

There are 2 types of schedulers defined at the moment
(1) RandomChoice
(2) LeastLoaded

The Chassis (add/delete/update) event is used to schedule router
gateway ports which have not been bound to any chassis.

A future patch will add functional tests.

Change-Id: I83fb561cee642b90c4de41bb8e472f47fa807191
Related-Bug: 1514995