Fix provider configuration in Vagrant file
To be able to setup correctly each VM, the provider sections of the Vagrant
file have to start with '<vm-name>.vm.provider'. In the current Vagrant file,
the provider sections of all the VMs start with 'config.vm.provider', which
causes all the VMs to be configured with the parametes of the last provider
section in the file.

Change-Id: I34e412be9e7e757a61d9ef4cc61262ac039fcd71