Fix provider network setup
A recent DevStack change [1] broke the networking-ovn gate. The
change is being reverted [2] to unbreak the gate. However, the
change found a weakness in the OVN provider network setup. In
particular when Q_USE_PROVIDERNET_FOR_PUBLIC=True, the
networking-ovn DevStack plugin should setup the bridge and
bridge mapping like it does when Q_USE_PROVIDER_NETWORKING=True.

The vagrant deployment broke in a different way because
PHYSICAL_NETWORK was not set (which is used to set the
flat networks supported). To fix this, the vagrant deployment
will now use the networking-ovn DevStack plugin for basic
provider network setup.

[1] https://review.openstack.org/#/c/343072/
[2] https://review.openstack.org/#/c/345820/

Related-Bug: #1605423
Change-Id: I63dd83b2374071703d341ec9e8d4d28c11174683