Support a mixed DPDK and non-DPDK environment
ovn-controller stores the datapath type and supported interface
types of the host in the external_ids column of the chassis table.

When the bind_port() function is called, this patch determines the
vif_type by reading these values from the chassis row where the port is to
be bound.

If the value of datapath type is 'netdev' and has 'dpdkvhostuser' in
the supported interface types,  vif_type will be set to 'vhostuser',
else vif_type will be set to 'ovs'.

The config param 'vif_type' is deprecated and no longer used.

Co-Authored-By: Richard Theis <rtheis@us.ibm.com>

Change-Id: Ia97198a5bceb5d8d39d8556346f3709846f3c5ad
Closes-bug: #1606458