Functional Test: Sync ACLs
Enhance functional tests to ensure that ACLs are properly
synchronized during OVN NB DB synchronization.

While enhancing the tests, a bug was discovered that resulted
in synchronization removing all ACLs for a port when it only
needed to create the port in OVN (ACLs already exist). This
was caused by duplicate ACLs being created during the port sync.

Change-Id: Ief64f73511de70ff16aaf2a2a9d26ae5a718c3b3
Closes-Bug: #1594241