Avoid error when removing addresses from address set
There are race conditions [1] that can trigger address removal from
address set, current code will result in a python exception:

   ValueError: list.remove(x): x not in list.

Even worse, if address set is left inconsistent with neutron DB
because of [1], e.g. an IP assigned to a port in neutron but not yet
reflected in OVN address set, then future updates of IPs to that
port will always fail on OVN side due to the exception above.

[1] https://bugs.launchpad.net/networking-ovn/+bug/1605089

Related-Bug: 1605089

Change-Id: Id002594a17f410aea3f6ada4ddabc418d64ca63d