Add initial OVS DB connection retry support
If the ML2 driver fails to initialize the OVS DB connections,
neutron server must be restarted before any OVN NB or SB transactions
can be processed. Effectively, the neutron server is half-dead. It is
able to process read-only requests but fails on all others. And the
error messages seen on requests and in the logs are often misleading.

This patch set adds initial OVS DB connection retry support and provides
more details on the error messages. With this patch, the neutron server
won't respond to any requests until the initial OVS DB connections can
be established. The connection attempts will be retried forever with an
exponential backoff up to a 3 minute wait between retries.

Also the retrying package was added to the requirements. It should have
been there already since it is used in ovsdb_monitor.py.

Note: OVS DB connection failures after initialization are retried
already.

Change-Id: I627b8c8ec0ff972e45bcd497c69f0fc94fbe6821
Closes-Bug: #1612435