Accept IPv6 RAs on the interface with the default route
Because the native services gate job sets ipv6 forwarding settings,
we stop accepting RAs from IPv6-only host environments. This leads
to a loss of external connectivity, which is bad for zuul running
tests and stuff.

Setting accept_ra to 2 will cause the RAs to be accepted.

This fix is based on [1].

[1] https://review.openstack.org/#/c/359490/

Depends-On: I1f30b550287e2ae0d49accee81a7ce4bddd3e0a4
Change-Id: Iae1b92ec9de087b6786dee372add064c54a01369