BugFix: L3 scheduler scheduling distributed routers
The L3 gateway router scheduler auto-schedules those routers
which have not been assigned a chassis (yet), when a new
chassis appears in the OVN_Southbound DB. This auto-scheduling
should only be done for those routers which are gateway routers.

Gateway Routers were either assigned a valid chassis during
creation or their options field was marked with
"chassis:neutron-ovn-invalid-chassis"

TODO: A unit test does need to be added for this case. It
will be part of series of ovsdb unit test case for the L3
scheduler logic (since the logic is based on the data in
the ovsdb).

Change-Id: I6e723e5394b564227db3f3834e29343197f4d86a
Closes-Bug: 1618508