Refuse port binding if not supported for host
Refuse port binding if not supported for host. The OVN chassis
for the host must exist and must have a bridge mapping for the
physical network when using flat or vlan networks. A message
will be issued for these cases since they indicate possible
deployment or configuration problems.

With this change, provisioning blocks must only be inserted
when there is an OVN chassis for the host.

Change-Id: I938441e97e12ccf441e0abd8d32c9f1e6a178d3a
Partially-Implements: blueprint routed-networks