Update port provisioning block registration
Deployments with the neutron DHCP agent must ensure that port status
is not changed to ACTIVE until OVN finishes provisioning the port.
This is done by adding provisioning blocks when a port is created
and when OVN marks the port down. The later provisioning block is
currently missing. Without the block, the neutron DHCP agent may
change the port status to ACTIVE before OVN is ready. In addition,
the provisioning block added during port creation should be done
during the ML2 precommit stage to avoid a similar problem.

Missing provisioning blocks creates a couple problems. The first
and easiest to notice problem is that shutting down an instance
keeps the port status ACTIVE when it should be DOWN. The second
and much more difficult problem is a timing window which is
generated during instance actions. The timing window can prevent
neutron from generating a network-vif-plugged event when nova
expects one. The event isn't generated because networking-ovn
didn't add a provisioning block to prevent the neutron DHCP
agent from interferring with VIF plugging.

Change-Id: I8355c94dbafaf7bc3f12871b7127db0087b67761
Closes-Bug: 1624120