Drop L3 agent from CI jobs.
Now that native SNAT and floating IP support has been merged for
networking-ovn, we can drop support for the Neutron L3 agent.  This
patch affects the following CI jobs:

tempest-dsvm-networking-ovn - Change to use OVN L3.  We will keep this
job because we still need a second temptest job that ensures the DHCP
agent still works, even though it's currently disabled due to a bug that
came up a few days ago.

rally-dsvm-networking-ovn - Use OVN L3 instead of the Neutron L3 agent.

The changes to the test regex are mostly to reduce the diff from this
file to devstackgatenativeservicesrc.

Partial-bug: #1626717
Change-Id: I16b337b64b7d96486ba0edeea1bb51151f0b8825
Signed-off-by: Russell Bryant <rbryant@redhat.com>