Update GW SNAT rule in if already exists for 'dnat_and_snat' type
When a floating ip is associated to a fixed ip, a row in NAT table
is created for the gateway router. When the port with the fixed ip
is deleted without disassociating the fip, the row still remains.
When the same floating ip is associated to a different fixed ip,
a new row is created in the NAT table. This causes the floating ip
unreachable.

In this patch, we first check for the presence of the NAT rule
for the floating ip and if present update the same to the new
fixed ip value.

Change-Id: Id36436032a37e0ae4c6ae9a23b1de5bf4da5b3fd
Closes-bug: #1642898