Fix issue of adding router interface from dashboard with specified IP
When adding a router-interface from dashboard with subnet and a
specified IP, the port in neutron DB contains security_groups default.
The networking-ovn adds router-interface IP to "Address_Set" of
Northbound-DB.
Actually, the router port need no security groups, this issue affects
lrouter's forwarding.
This patch fixes this bug. If security groups are present in a trusted
port we ignore them.

Change-Id: Ie872510416f50fcee7ec6d3760006b095621bff9
Closes-Bug: #1647949
Signed-off-by: Dong Jun <dongj@dtdream.com>