Sync devstackgaterc with devstackgatenativeservicesrc.
The main tempest job and the "native-services" job have been running the
same configuration for a while.  At one point we had thought that we
might want to retain support for the DHCP agent, but I propose that we
drop it and revisit if there's demand for it in the future.

After this patch lands, I'm going to remove references to the
nativeservicesrc file in CI config and then we can remove the file from
our repo.

Change-Id: I3aa7a6ac10e4da00d5830f68d7c9d3d684f91218
Closes-bug: 1639806