Refresh devstack testing docs.
One big difference is that we create both a public and private network
by default.  The previous docs only showed a private network.

Update most commands to use the openstack client instead of the older
project specific clients.

Add a section on allocating a floating IP address so that you can
connect to the VMs.

Change-Id: I48850d87db1843fcea0ff00074c6cd17980a39f1
Signed-off-by: Russell Bryant <rbryant@redhat.com>