Implement disassociate_floatingips in l3_ovn.py
When VM with floating IP associated was deleted, dnat_and_snat row in
OVN_Northbound DB was not removed. Then delete this floating IP,
dnat_and_snat is still in OVN DB.
Currently a workaround in _update_floating_ip_in_ovn is, when
re-associating, getting the list of nat rules and checking if the
external_ip with type 'dnat_and_snat' already exists or not. If exists,
set the new value.
However, disassociate_floatingips is the conventional approach for
l3 plugin to handle the callback of VM port' deleting, this patch
implements the function in l3_ovn plugin.

Change-Id: Ib487c00905d08e1f47f77bd4ddb2810c8ee19ea9
Closes-Bug: #1677581
Signed-off-by: Dong Jun <dongj@dtdream.com>