Update local.conf.sample to enable automatic host discovery for cell
Cell is enabled by default since Ocata. Because of the change, when
a new compute node is added the host needs to be discovered by the
cell. Otherwise, there will be nova scheduler error when booting a
vm: host xxx is not mapped to any cell. Discovery can be done either
manually or by enabling automatic discovery. This patch is to enable
the automatic discovery, to avoid unexpected error when following the
steps in [1].

[1] https://docs.openstack.org/developer/networking-ovn/testing.html

Change-Id: I1719fc8e3d0a0e6b5a003a73087886df18fc08ea