Allow tuning the probe interval from the IDL session
This patch is adding a new configuration options called
"ovsdb_probe_interval" which is used to tune the probe_interval attribute
from OVS python library. There are some use-cases (e.g load testings)
where people might want to disable or at least set a less agressive
value for that probing. Based on previous discussions, by default,
networking-ovn will disable probing. The overhead causing by agressive
probing overweight the benefits of it.

Also, a FIXME note was left in the code because OVS doesn't expose a
public method to configure the probing in the Idl class (which we
inherint from in networking-ovn), so for now we need to make use of a
private attribute which is not ideal. A patch for OVS to expose these
configuration has been submitted already [0]

[0]
https://mail.openvswitch.org/pipermail/ovs-dev/2017-April/330531.html

Change-Id: Ia77bbea19b176ecead4b4aed3674b59a49ad06c4
Closes-Bug: #1680146