Call the ovs 'Stream' class ssl_set_* functions only if required
The present code calls the Stream.ssl_set_* functions irrespective
of the SSL file options set by the user or not.

This patch calls these functions only if required. With this patch,
we can use python ovs lib < 2.7 without any issues. If the requirement
is to use SSL for ovsdb connection, the user has to use python ovs
lib >= 2.7.

Change-Id: I7f4ec0015591b2891308147f3645d8ce674158dd