support distributed floating ip
User can enable the distributed floating ip by configuring the
option enable_distributed_floating_ip.

This require the user to configure the physical network map
to connecting the floating ip's network on each compute node.

the Nat action for floating IP will be done locally, not
gateway. This will saves the path to the external network.

Change-Id: I4e8d6811b5a544e68b184fe457391b35e4516d01
Co-authored-by: Dong Jun <dongj@dtdream.com>