Inherit from ovsdbapp API classes
This will automatically give nested transaction support and will
make it easier when moving this code into ovsdbapp.

The OvsdbSbOvnIdl class didn't currently have the ability to
create transactions, presumably because the code currently treats
the SB DB as read-only. In the test code, though, many hoops are
jumped through to create a "fake" API so that a Chassis can be
added for test purposes. In a generic implementation of a SB API,
read and write operations will be added, so we might as well go
ahead and add this now.

Change-Id: I9820ef9de7a35a8287e1c14c122d9584c9c1e8cf