Remove two log traces that are no longer useful
This patch removes two traces that used to be useful during
development to detect when a process had the lock to handle a
notification. Since IDL will call notify() for every "update" rpc
method it receives from ovsdb-server, this trace can be quite frequent
and noisy.

In a normal tempest execution in gate it has around a 7% (~6500
occurences) of the total log entries.

Change-Id: I4d2a0be86f07b55d631bacf21d090635625a8ec6