call provisioning_complete conditionally
Right now, we are calling provisioning_complete without checking
if the resource is added or not in provisioning block and each
provisioning_complete is triggering port's AFTER_UPDATE
event.

Unit test 'test_registry_notify_before_after_port_binding',
checks number of AFTER_UPDATE events, and is expecting only 2
events when port is updated. Neutron emits only 2 events.
networking-ovn triggering this third AFTER_UPDATE through
provisioning_complete. As neutron or networking-ovn is not
adding port in provisioning block, this provisioning_complete
is not required. So in this patch, we call provisioning_complete
only if the port is added in provisioning block.

Closes-bug: #1707215
Change-Id: Icab5514e29d1bd9927dbbdc68def73eff6be22ce