Fix gate issues
Right now, our gate is broken due to bug 1707215. Until it's fixed
this patch skips this test so that CI passes again.

Also fix the functional test failure in test_ovn_db_sync by using
ovs 2.7 branch. The reason for the failure is because, recently
ACL table was enhanced to add 2 new columns - name and severity.
If we use these new columns in the acl code AS IS,
'gate-tempest-dsvm-networking-ovn-ovs-release' will be broken.
We need to handle the DB changes properly in the networking-ovn code
in order to support master and latest release branch code of OVS.

Related-Bug: #1707215
Closes-bug: #1708594
Change-Id: Ibc0f5fbde8bba6dd264a2653d515bac4bfc9a8a4