Drop journaling related code
This patch is removing the code related to journaling. A new approach
to solve the database consistency problem has been proposed and accepted
at e34e5acf27a86934bbb8ee47e59806c617099870.

The skeleton for migration (and tests) have been kept in the patch
because that will be used later by the new solution.

Change-Id: Id036ca27df6117e8317238bc719ab21f49c431fd